scapper/requirements.txt
pymongo
selenium
beautifulsoup4
lxml
python-dotenv
dnspython
pycryptodome